+++ miden-crypto/src/hash/mod.rs
@@ -3,7 +3,7 @@
 use crate::{Felt, Word, ZERO};
 
 /// Generic digest types for binary hash functions.
-pub mod digest;
+pub(crate) mod digest;
 
 /// Blake3 hash function.
 pub mod blake;